requirements-dev.txt
ansible==2.7.7
boto3==1.6.11
black
flake8
ipdb
isort
mock==2.0.0
pytest==4.3.0
pytest-django==3.4.8
